factory_management/requirements.txt
Flask
Flask-SQLAlchemy
pyjwt
Flask-Migrate
